Refactored KMeansParallel to allow it to take starting centers and starting costs.

+++ src/main/scala/com/massivedatascience/clusterer/KMeansPlusPlus.scala
@@ -49,7 +49,7 @@ class KMeansPlusPlus(ops: BregmanPointOps) extends Serializable with SparkHelper
    */
 
   def getCenters(
-    seed: Int,
+    seed: Long,
     candidateCenters: Array[BregmanCenter],
     weights: Array[Double],
     k: Int,
